Updated: Sort the folder.
