+++ libft/srcs/mini_ls/ft_get_folder.c
@@ -6,7 +6,7 @@
 /*   By: alerandy <alerandy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/13 11:35:10 by alerandy          #+#    #+#             */
-/*   Updated: 2019/08/10 15:44:29 by alerandy         ###   ########.fr       */
+/*   Updated: 2019/08/11 00:33:23 by alerandy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,17 +30,6 @@ static inline int		count_file(char *path)
 	return (i);
 }
 
-void					ft_free_folder(t_ls_folder **folder)
-{
-	int		i;
-
-	i = -1;
-	while ((*folder)->files[++i])
-		ft_strdel(&((*folder)->files[i]));
-	ft_memdel((void**)&(*folder)->files);
-	ft_memdel((void**)folder);
-}
-
 void					ft_refresh_folder(t_ls_folder **folder)
 {
 	char	*path;
@@ -53,6 +42,11 @@ void					ft_refresh_folder(t_ls_folder **folder)
 	ft_strdel(&path);
 }
 
+static inline int		voidcmp(void *s1, void *s2)
+{
+	return (ft_strcmp((char*)s1, (char*)s2));
+}
+
 static inline void		fill_folder(t_ls_folder *folder, DIR *dir, char *path)
 {
 	_DIRENT	*fold;
@@ -63,6 +57,7 @@ static inline void		fill_folder(t_ls_folder *folder, DIR *dir, char *path)
 					? ft_strjoin(path, "/") : ft_strdup(path);
 	while ((fold = readdir(dir)))
 		(folder->files)[i++] = ft_strdup(fold->d_name);
+	ft_sortarray((void**)folder->files, folder->files_amount, &voidcmp);
 	closedir(dir);
 }
